+++ impeller/entity/contents/gradient_generator.cc
@@ -42,4 +42,15 @@ std::shared_ptr<Texture> CreateGradientTexture(
   return texture;
 }
 
+std::vector<StopData> CreateGradientColors(const std::vector<Color>& colors,
+                                           const std::vector<Scalar>& stops) {
+  FML_DCHECK(stops.size() == colors.size());
+
+  std::vector<StopData> result(stops.size());
+  for (auto i = 0u; i < stops.size(); i++) {
+    result[i] = {.color = colors[i], .stop = stops[i]};
+  }
+  return result;
+}
+
 }  // namespace impeller
