+++ impeller/entity/contents/gradient_generator.h
@@ -14,6 +14,7 @@
 #include "impeller/geometry/gradient.h"
 #include "impeller/geometry/path.h"
 #include "impeller/geometry/point.h"
+#include "impeller/renderer/shader_types.h"
 
 namespace impeller {
 
@@ -26,5 +27,26 @@ class Context;
 std::shared_ptr<Texture> CreateGradientTexture(
     const GradientData& gradient_data,
     const std::shared_ptr<impeller::Context>& context);
+
+struct StopData {
+  Color color;
+  Scalar stop;
+  Padding<12> _padding_;
+};
+
+/**
+ * @brief Populate a vector with the interpolated colors for the linear gradient
+ * described  by colors and stops.
+ *
+ * If the returned result is std::nullopt, the original color buffer can be used
+ * instead.
+ *
+ * @param colors
+ * @param stops
+ * @return StopData
+ */
+std::vector<StopData> CreateGradientColors(const std::vector<Color>& colors,
+                                           const std::vector<Scalar>& stops);
+
 
 }  // namespace impeller
