+++ impeller/geometry/gradient.cc
@@ -98,15 +98,4 @@ GradientData CreateGradientBuffer(const std::vector<Color>& colors,
   return data;
 }
 
-std::vector<StopData> CreateGradientColors(const std::vector<Color>& colors,
-                                           const std::vector<Scalar>& stops) {
-  FML_DCHECK(stops.size() == colors.size());
-
-  std::vector<StopData> result(stops.size());
-  for (auto i = 0u; i < stops.size(); i++) {
-    result[i] = {.color = colors[i], .stop = stops[i]};
-  }
-  return result;
-}
-
 }  // namespace impeller
