+++ impeller/geometry/gradient.h
@@ -10,7 +10,6 @@
 #include "impeller/geometry/color.h"
 #include "impeller/geometry/path.h"
 #include "impeller/geometry/point.h"
-#include "impeller/renderer/shader_types.h"
 
 namespace impeller {
 
@@ -30,25 +29,5 @@ struct GradientData {
  */
 GradientData CreateGradientBuffer(const std::vector<Color>& colors,
                                   const std::vector<Scalar>& stops);
-
-struct StopData {
-  Color color;
-  Scalar stop;
-  Padding<12> _padding_;
-};
-
-/**
- * @brief Populate a vector with the interpolated colors for the linear gradient
- * described  by colors and stops.
- *
- * If the returned result is std::nullopt, the original color buffer can be used
- * instead.
- *
- * @param colors
- * @param stops
- * @return StopData
- */
-std::vector<StopData> CreateGradientColors(const std::vector<Color>& colors,
-                                           const std::vector<Scalar>& stops);
 
 }  // namespace impeller
